fs9922: fix digits in analog payload
The properties sr_analog_encoding.digits and sr_analog_spec.spec_digits
expect the number of decimal places (if positive) of the measured value in
relation to the available digits (4 digits for the fs9922). This information
is used by software based on libsigrok to get a representation of the
measured value as close as possible to the display of the DMM itself.

This commit also changes some variable names to better reflect their usage.