HACK tests: add DMM packet parser unit tests, start with bm25x
Strictly speaking the tests/ hierarchy is optional, depends on the
'check' library, and exclusively accesses the library's public API,
like a regular application would (and finds a stable API because it
only covers the public part).

Yet coverage for DMM packet parsers is desirable. Implement a public
suite_dmm_packets() routine within the library, which gets registered
with the external test application, and internally accesses library's
DMM packets parse() routines. This increases test coverage while the
library API is only minimally affected. The (optional) presence of
test code in the resulting library is considered acceptable.

TODO This was only build and run tested with libcheck 0.10. Absence of
the optional dependency is yet to get checked. As is 0.11 compatibilty
(ck_assert_float_*() helpers). Frank's copyright is assumed in the main
source file, because the test implementation is modelled after his
development branch.