PoC: Add default implementation for r/w mutex
This commit adds a default implemetation for a read/write mutex to make
the device communication thread safe.

This implementaion protects access to the `config_get()` and the `config_set()`
functions and also to the via `dev_acquisition_start()` registered callback
function by a mutex.

If a driver has to implement its own shared access to the devices resources
(like every scpi driber or the arachnid-labs-re-load-pro drver) this can be
done by disabling the default behaviour by setting `disable_default_mutex` of
`sr_dev_inst` to `TRUE` in the drivers `scan` function.